requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pandas==0.25.3
opencv-python==4.1.1.26
pyparsing==2.4.4
python-dateutil==2.8.1
six==1.13.0
torch==1.3.1
Pillow==6.2.1
torchvision==0.4.2
